Revert "[in_app_purchase_storekit] Add support for purchase and transactions …"
This reverts commit c1435edc9c5444eb255ca18f1e879fc7a02d0959.
