+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,7 +1,3 @@
-## 0.3.18+2
-
-* Adds support for StoreKit2's `purchase` and `transactions`
-
 ## 0.3.18+1
 
 * Adds support for StoreKit2's `canMakePayments` and `products`
