+++ packages/in_app_purchase/in_app_purchase_storekit/darwin/Classes/StoreKit2/sk2_pigeon.g.swift
@@ -58,12 +58,6 @@ private func wrapError(_ error: Any) -> [Any?] {
   ]
 }
 
-private func createConnectionError(withChannelName channelName: String) -> PigeonError {
-  return PigeonError(
-    code: "channel-error", message: "Unable to establish connection on channel: '\(channelName)'.",
-    details: "")
-}
-
 private func isNullish(_ value: Any?) -> Bool {
   return value is NSNull || value == nil
 }
@@ -100,12 +94,6 @@ enum SK2SubscriptionPeriodUnitMessage: Int {
   case week = 1
   case month = 2
   case year = 3
-}
-
-enum SK2ProductPurchaseResultMessage: Int {
-  case success = 0
-  case userCancelled = 1
-  case pending = 2
 }
 
 /// Generated class from Pigeon that represents data sent in messages.
@@ -287,111 +275,6 @@ struct SK2PriceLocaleMessage {
   }
 }
 
-/// Generated class from Pigeon that represents data sent in messages.
-struct SK2ProductPurchaseOptionsMessage {
-  var appAccountToken: String? = nil
-  var quantity: Int64? = nil
-
-  // swift-format-ignore: AlwaysUseLowerCamelCase
-  static func fromList(_ pigeonVar_list: [Any?]) -> SK2ProductPurchaseOptionsMessage? {
-    let appAccountToken: String? = nilOrValue(pigeonVar_list[0])
-    let quantity: Int64? =
-      isNullish(pigeonVar_list[1])
-      ? nil
-      : (pigeonVar_list[1] is Int64?
-        ? pigeonVar_list[1] as! Int64? : Int64(pigeonVar_list[1] as! Int32))
-
-    return SK2ProductPurchaseOptionsMessage(
-      appAccountToken: appAccountToken,
-      quantity: quantity
-    )
-  }
-  func toList() -> [Any?] {
-    return [
-      appAccountToken,
-      quantity,
-    ]
-  }
-}
-
-/// Generated class from Pigeon that represents data sent in messages.
-struct SK2TransactionMessage {
-  var id: Int64
-  var originalId: Int64
-  var productId: String
-  var purchaseDate: String
-  var purchasedQuantity: Int64
-  var appAccountToken: String? = nil
-  var restoring: Bool
-  var error: SK2ErrorMessage? = nil
-
-  // swift-format-ignore: AlwaysUseLowerCamelCase
-  static func fromList(_ pigeonVar_list: [Any?]) -> SK2TransactionMessage? {
-    let id =
-      pigeonVar_list[0] is Int64 ? pigeonVar_list[0] as! Int64 : Int64(pigeonVar_list[0] as! Int32)
-    let originalId =
-      pigeonVar_list[1] is Int64 ? pigeonVar_list[1] as! Int64 : Int64(pigeonVar_list[1] as! Int32)
-    let productId = pigeonVar_list[2] as! String
-    let purchaseDate = pigeonVar_list[3] as! String
-    let purchasedQuantity =
-      pigeonVar_list[4] is Int64 ? pigeonVar_list[4] as! Int64 : Int64(pigeonVar_list[4] as! Int32)
-    let appAccountToken: String? = nilOrValue(pigeonVar_list[5])
-    let restoring = pigeonVar_list[6] as! Bool
-    let error: SK2ErrorMessage? = nilOrValue(pigeonVar_list[7])
-
-    return SK2TransactionMessage(
-      id: id,
-      originalId: originalId,
-      productId: productId,
-      purchaseDate: purchaseDate,
-      purchasedQuantity: purchasedQuantity,
-      appAccountToken: appAccountToken,
-      restoring: restoring,
-      error: error
-    )
-  }
-  func toList() -> [Any?] {
-    return [
-      id,
-      originalId,
-      productId,
-      purchaseDate,
-      purchasedQuantity,
-      appAccountToken,
-      restoring,
-      error,
-    ]
-  }
-}
-
-/// Generated class from Pigeon that represents data sent in messages.
-struct SK2ErrorMessage {
-  var code: Int64
-  var domain: String
-  var userInfo: [String?: Any?]? = nil
-
-  // swift-format-ignore: AlwaysUseLowerCamelCase
-  static func fromList(_ pigeonVar_list: [Any?]) -> SK2ErrorMessage? {
-    let code =
-      pigeonVar_list[0] is Int64 ? pigeonVar_list[0] as! Int64 : Int64(pigeonVar_list[0] as! Int32)
-    let domain = pigeonVar_list[1] as! String
-    let userInfo: [String?: Any?]? = nilOrValue(pigeonVar_list[2])
-
-    return SK2ErrorMessage(
-      code: code,
-      domain: domain,
-      userInfo: userInfo
-    )
-  }
-  func toList() -> [Any?] {
-    return [
-      code,
-      domain,
-      userInfo,
-    ]
-  }
-}
-
 private class sk2_pigeonPigeonCodecReader: FlutterStandardReader {
   override func readValue(ofType type: UInt8) -> Any? {
     switch type {
@@ -420,27 +303,15 @@ private class sk2_pigeonPigeonCodecReader: FlutterStandardReader {
       }
       return nil
     case 133:
-      let enumResultAsInt: Int? = nilOrValue(self.readValue() as? Int)
-      if let enumResultAsInt = enumResultAsInt {
-        return SK2ProductPurchaseResultMessage(rawValue: enumResultAsInt)
-      }
-      return nil
-    case 134:
       return SK2SubscriptionOfferMessage.fromList(self.readValue() as! [Any?])
-    case 135:
+    case 134:
       return SK2SubscriptionPeriodMessage.fromList(self.readValue() as! [Any?])
-    case 136:
+    case 135:
       return SK2SubscriptionInfoMessage.fromList(self.readValue() as! [Any?])
-    case 137:
+    case 136:
       return SK2ProductMessage.fromList(self.readValue() as! [Any?])
-    case 138:
+    case 137:
       return SK2PriceLocaleMessage.fromList(self.readValue() as! [Any?])
-    case 139:
-      return SK2ProductPurchaseOptionsMessage.fromList(self.readValue() as! [Any?])
-    case 140:
-      return SK2TransactionMessage.fromList(self.readValue() as! [Any?])
-    case 141:
-      return SK2ErrorMessage.fromList(self.readValue() as! [Any?])
     default:
       return super.readValue(ofType: type)
     }
@@ -461,32 +332,20 @@ private class sk2_pigeonPigeonCodecWriter: FlutterStandardWriter {
     } else if let value = value as? SK2SubscriptionPeriodUnitMessage {
       super.writeByte(132)
       super.writeValue(value.rawValue)
-    } else if let value = value as? SK2ProductPurchaseResultMessage {
-      super.writeByte(133)
-      super.writeValue(value.rawValue)
     } else if let value = value as? SK2SubscriptionOfferMessage {
-      super.writeByte(134)
+      super.writeByte(133)
       super.writeValue(value.toList())
     } else if let value = value as? SK2SubscriptionPeriodMessage {
-      super.writeByte(135)
+      super.writeByte(134)
       super.writeValue(value.toList())
     } else if let value = value as? SK2SubscriptionInfoMessage {
-      super.writeByte(136)
+      super.writeByte(135)
       super.writeValue(value.toList())
     } else if let value = value as? SK2ProductMessage {
-      super.writeByte(137)
+      super.writeByte(136)
       super.writeValue(value.toList())
     } else if let value = value as? SK2PriceLocaleMessage {
-      super.writeByte(138)
-      super.writeValue(value.toList())
-    } else if let value = value as? SK2ProductPurchaseOptionsMessage {
-      super.writeByte(139)
-      super.writeValue(value.toList())
-    } else if let value = value as? SK2TransactionMessage {
-      super.writeByte(140)
-      super.writeValue(value.toList())
-    } else if let value = value as? SK2ErrorMessage {
-      super.writeByte(141)
+      super.writeByte(137)
       super.writeValue(value.toList())
     } else {
       super.writeValue(value)
@@ -513,13 +372,6 @@ protocol InAppPurchase2API {
   func canMakePayments() throws -> Bool
   func products(
     identifiers: [String], completion: @escaping (Result<[SK2ProductMessage], Error>) -> Void)
-  func purchase(
-    id: String, options: SK2ProductPurchaseOptionsMessage?,
-    completion: @escaping (Result<SK2ProductPurchaseResultMessage, Error>) -> Void)
-  func transactions(completion: @escaping (Result<[SK2TransactionMessage], Error>) -> Void)
-  func finish(id: Int64, completion: @escaping (Result<Void, Error>) -> Void)
-  func startListeningToTransactions() throws
-  func stopListeningToTransactions() throws
 }
 
 /// Generated setup class from Pigeon to handle messages through the `binaryMessenger`.
@@ -566,136 +418,6 @@ class InAppPurchase2APISetup {
       }
     } else {
       productsChannel.setMessageHandler(nil)
-    }
-    let purchaseChannel = FlutterBasicMessageChannel(
-      name:
-        "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.purchase\(channelSuffix)",
-      binaryMessenger: binaryMessenger, codec: codec)
-    if let api = api {
-      purchaseChannel.setMessageHandler { message, reply in
-        let args = message as! [Any?]
-        let idArg = args[0] as! String
-        let optionsArg: SK2ProductPurchaseOptionsMessage? = nilOrValue(args[1])
-        api.purchase(id: idArg, options: optionsArg) { result in
-          switch result {
-          case .success(let res):
-            reply(wrapResult(res))
-          case .failure(let error):
-            reply(wrapError(error))
-          }
-        }
-      }
-    } else {
-      purchaseChannel.setMessageHandler(nil)
-    }
-    let transactionsChannel = FlutterBasicMessageChannel(
-      name:
-        "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.transactions\(channelSuffix)",
-      binaryMessenger: binaryMessenger, codec: codec)
-    if let api = api {
-      transactionsChannel.setMessageHandler { _, reply in
-        api.transactions { result in
-          switch result {
-          case .success(let res):
-            reply(wrapResult(res))
-          case .failure(let error):
-            reply(wrapError(error))
-          }
-        }
-      }
-    } else {
-      transactionsChannel.setMessageHandler(nil)
-    }
-    let finishChannel = FlutterBasicMessageChannel(
-      name: "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.finish\(channelSuffix)",
-      binaryMessenger: binaryMessenger, codec: codec)
-    if let api = api {
-      finishChannel.setMessageHandler { message, reply in
-        let args = message as! [Any?]
-        let idArg = args[0] is Int64 ? args[0] as! Int64 : Int64(args[0] as! Int32)
-        api.finish(id: idArg) { result in
-          switch result {
-          case .success:
-            reply(wrapResult(nil))
-          case .failure(let error):
-            reply(wrapError(error))
-          }
-        }
-      }
-    } else {
-      finishChannel.setMessageHandler(nil)
-    }
-    let startListeningToTransactionsChannel = FlutterBasicMessageChannel(
-      name:
-        "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.startListeningToTransactions\(channelSuffix)",
-      binaryMessenger: binaryMessenger, codec: codec)
-    if let api = api {
-      startListeningToTransactionsChannel.setMessageHandler { _, reply in
-        do {
-          try api.startListeningToTransactions()
-          reply(wrapResult(nil))
-        } catch {
-          reply(wrapError(error))
-        }
-      }
-    } else {
-      startListeningToTransactionsChannel.setMessageHandler(nil)
-    }
-    let stopListeningToTransactionsChannel = FlutterBasicMessageChannel(
-      name:
-        "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.stopListeningToTransactions\(channelSuffix)",
-      binaryMessenger: binaryMessenger, codec: codec)
-    if let api = api {
-      stopListeningToTransactionsChannel.setMessageHandler { _, reply in
-        do {
-          try api.stopListeningToTransactions()
-          reply(wrapResult(nil))
-        } catch {
-          reply(wrapError(error))
-        }
-      }
-    } else {
-      stopListeningToTransactionsChannel.setMessageHandler(nil)
-    }
-  }
-}
-/// Generated protocol from Pigeon that represents Flutter messages that can be called from Swift.
-protocol InAppPurchase2CallbackAPIProtocol {
-  func onTransactionsUpdated(
-    newTransaction newTransactionArg: SK2TransactionMessage,
-    completion: @escaping (Result<Void, PigeonError>) -> Void)
-}
-class InAppPurchase2CallbackAPI: InAppPurchase2CallbackAPIProtocol {
-  private let binaryMessenger: FlutterBinaryMessenger
-  private let messageChannelSuffix: String
-  init(binaryMessenger: FlutterBinaryMessenger, messageChannelSuffix: String = "") {
-    self.binaryMessenger = binaryMessenger
-    self.messageChannelSuffix = messageChannelSuffix.count > 0 ? ".\(messageChannelSuffix)" : ""
-  }
-  var codec: sk2_pigeonPigeonCodec {
-    return sk2_pigeonPigeonCodec.shared
-  }
-  func onTransactionsUpdated(
-    newTransaction newTransactionArg: SK2TransactionMessage,
-    completion: @escaping (Result<Void, PigeonError>) -> Void
-  ) {
-    let channelName: String =
-      "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2CallbackAPI.onTransactionsUpdated\(messageChannelSuffix)"
-    let channel = FlutterBasicMessageChannel(
-      name: channelName, binaryMessenger: binaryMessenger, codec: codec)
-    channel.sendMessage([newTransactionArg] as [Any?]) { response in
-      guard let listResponse = response as? [Any?] else {
-        completion(.failure(createConnectionError(withChannelName: channelName)))
-        return
-      }
-      if listResponse.count > 1 {
-        let code: String = listResponse[0] as! String
-        let message: String? = nilOrValue(listResponse[1])
-        let details: String? = nilOrValue(listResponse[2])
-        completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else {
-        completion(.success(Void()))
-      }
     }
   }
 }
