+++ packages/in_app_purchase/in_app_purchase_storekit/example/ios/Runner.xcodeproj/xcshareddata/xcschemes/Runner.xcscheme
@@ -39,7 +39,6 @@
       </MacroExpansion>
       <Testables>
          <TestableReference
-            parallelizable = "NO">
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
