+++ packages/in_app_purchase/in_app_purchase_storekit/example/ios/Runner/Configuration.storekit
@@ -43,21 +43,6 @@
       "productID" : "upgrade",
       "referenceName" : "upgrade",
       "type" : "NonConsumable"
-    },
-    {
-      "displayPrice" : "0.49",
-      "familyShareable" : false,
-      "internalID" : "491F25B6",
-      "localizations" : [
-        {
-          "description" : "A consumable product.",
-          "displayName" : "Consumable",
-          "locale" : "en_US"
-        }
-      ],
-      "productID" : "consumable_discounted",
-      "referenceName" : "consumable_discounted",
-      "type" : "Consumable"
     }
   ],
   "settings" : {
@@ -74,10 +59,7 @@
         "name" : "Load Products"
       },
       {
-        "current" : {
-          "index" : 1,
-          "type" : "purchase"
-        },
+        "current" : null,
         "enabled" : false,
         "name" : "Purchase"
       },
@@ -92,10 +74,7 @@
         "name" : "App Store Sync"
       },
       {
-        "current" : {
-          "index" : 2,
-          "type" : "generic"
-        },
+        "current" : null,
         "enabled" : false,
         "name" : "Subscription Status"
       },
@@ -127,7 +106,7 @@
       "localizations" : [
 
       ],
-      "name" : "Silver Subscriptions",
+      "name" : "Example Subscriptions",
       "subscriptions" : [
         {
           "adHocOffers" : [
@@ -156,16 +135,7 @@
           "winbackOffers" : [
 
           ]
-        }
-      ]
-    },
-    {
-      "id" : "5D664105",
-      "localizations" : [
-
-      ],
-      "name" : "Gold Subscription",
-      "subscriptions" : [
+        },
         {
           "adHocOffers" : [
 
@@ -175,8 +145,8 @@
           ],
           "displayPrice" : "5.99",
           "familyShareable" : false,
-          "groupNumber" : 1,
-          "internalID" : "12B805AE",
+          "groupNumber" : 2,
+          "internalID" : "0BC7FF5E",
           "introductoryOffer" : null,
           "localizations" : [
             {
@@ -188,49 +158,7 @@
           "productID" : "subscription_gold",
           "recurringSubscriptionPeriod" : "P1M",
           "referenceName" : "subscription_gold",
-          "subscriptionGroupID" : "5D664105",
-          "type" : "RecurringSubscription",
-          "winbackOffers" : [
-
-          ]
-        }
-      ]
-    },
-    {
-      "id" : "3B74452E",
-      "localizations" : [
-
-      ],
-      "name" : "Discounted Subscriptions",
-      "subscriptions" : [
-        {
-          "adHocOffers" : [
-
-          ],
-          "codeOffers" : [
-
-          ],
-          "displayPrice" : "0.99",
-          "familyShareable" : false,
-          "groupNumber" : 1,
-          "internalID" : "ECFBF793",
-          "introductoryOffer" : {
-            "displayPrice" : "0.99",
-            "internalID" : "62DAF06C",
-            "paymentMode" : "payUpFront",
-            "subscriptionPeriod" : "P1M"
-          },
-          "localizations" : [
-            {
-              "description" : "",
-              "displayName" : "",
-              "locale" : "en_US"
-            }
-          ],
-          "productID" : "subscription_discounted",
-          "recurringSubscriptionPeriod" : "P1M",
-          "referenceName" : "subscription_discounted",
-          "subscriptionGroupID" : "3B74452E",
+          "subscriptionGroupID" : "D0FEE8D8",
           "type" : "RecurringSubscription",
           "winbackOffers" : [
