+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/in_app_purchase_storekit_platform.dart
@@ -33,26 +33,16 @@ class InAppPurchaseStoreKitPlatform extends InAppPurchasePlatform {
   /// Experimental flag for StoreKit2.
   static bool _useStoreKit2 = false;
 
-  /// StoreKit1
   static late SKPaymentQueueWrapper _skPaymentQueueWrapper;
-  static late _TransactionObserver _sk1transactionObserver;
-
-  /// StoreKit2
-  static late SK2TransactionObserverWrapper _sk2transactionObserver;
+  static late _TransactionObserver _observer;
 
   @override
-  Stream<List<PurchaseDetails>> get purchaseStream => _useStoreKit2
-      ? _sk2transactionObserver.transactionsCreatedController.stream
-      : _sk1transactionObserver.purchaseUpdatedController.stream;
+  Stream<List<PurchaseDetails>> get purchaseStream =>
+      _observer.purchaseUpdatedController.stream;
 
   /// Callback handler for transaction status changes.
   @visibleForTesting
-  static SKTransactionObserverWrapper get observer => _sk1transactionObserver;
-
-  /// Callback handler for transaction status changes for StoreKit2 transactions
-  @visibleForTesting
-  static SK2TransactionObserverWrapper get sk2transactionObserver =>
-      _sk2transactionObserver;
+  static SKTransactionObserverWrapper get observer => _observer;
 
   /// Registers this class as the default instance of [InAppPurchasePlatform].
   static void registerPlatform() {
@@ -67,26 +57,15 @@ class InAppPurchaseStoreKitPlatform extends InAppPurchasePlatform {
 
     _skPaymentQueueWrapper = SKPaymentQueueWrapper();
 
-    if (_useStoreKit2) {
-      final StreamController<List<PurchaseDetails>> updateController2 =
-          StreamController<List<PurchaseDetails>>.broadcast(
-        onListen: () => SK2Transaction.startListeningToTransactions(),
-        onCancel: () => SK2Transaction.stopListeningToTransactions(),
-      );
-      _sk2transactionObserver = SK2TransactionObserverWrapper(
-          transactionsCreatedController: updateController2);
-      InAppPurchase2CallbackAPI.setUp(_sk2transactionObserver);
-    } else {
-      // Create a purchaseUpdatedController and notify the native side when to
-      // start of stop sending updates.
-      final StreamController<List<PurchaseDetails>> updateController =
-          StreamController<List<PurchaseDetails>>.broadcast(
-        onListen: () => _skPaymentQueueWrapper.startObservingTransactionQueue(),
-        onCancel: () => _skPaymentQueueWrapper.stopObservingTransactionQueue(),
-      );
-      _sk1transactionObserver = _TransactionObserver(updateController);
-      _skPaymentQueueWrapper.setTransactionObserver(observer);
-    }
+    // Create a purchaseUpdatedController and notify the native side when to
+    // start of stop sending updates.
+    final StreamController<List<PurchaseDetails>> updateController =
+        StreamController<List<PurchaseDetails>>.broadcast(
+      onListen: () => _skPaymentQueueWrapper.startObservingTransactionQueue(),
+      onCancel: () => _skPaymentQueueWrapper.stopObservingTransactionQueue(),
+    );
+    _observer = _TransactionObserver(updateController);
+    _skPaymentQueueWrapper.setTransactionObserver(observer);
   }
 
   @override
@@ -99,17 +78,6 @@ class InAppPurchaseStoreKitPlatform extends InAppPurchasePlatform {
 
   @override
   Future<bool> buyNonConsumable({required PurchaseParam purchaseParam}) async {
-    if (_useStoreKit2) {
-      final SK2ProductPurchaseOptions options = SK2ProductPurchaseOptions(
-          quantity: purchaseParam is AppStorePurchaseParam
-              ? purchaseParam.quantity
-              : 1,
-          appAccountToken: purchaseParam.applicationUserName);
-      await SK2Product.purchase(purchaseParam.productDetails.id,
-          options: options);
-
-      return true;
-    }
     await _skPaymentQueueWrapper.addPayment(SKPaymentWrapper(
         productIdentifier: purchaseParam.productDetails.id,
         quantity:
@@ -134,13 +102,9 @@ class InAppPurchaseStoreKitPlatform extends InAppPurchasePlatform {
   @override
   Future<void> completePurchase(PurchaseDetails purchase) {
     assert(
-      purchase is AppStorePurchaseDetails || purchase is SK2PurchaseDetails,
+      purchase is AppStorePurchaseDetails,
       'On iOS, the `purchase` should always be of type `AppStorePurchaseDetails`.',
     );
-
-    if (_useStoreKit2) {
-      return SK2Transaction.finish(int.parse(purchase.purchaseID!));
-    }
 
     return _skPaymentQueueWrapper.finishTransaction(
       (purchase as AppStorePurchaseDetails).skPaymentTransaction,
@@ -149,12 +113,11 @@ class InAppPurchaseStoreKitPlatform extends InAppPurchasePlatform {
 
   @override
   Future<void> restorePurchases({String? applicationUserName}) async {
-    return _sk1transactionObserver
+    return _observer
         .restoreTransactions(
             queue: _skPaymentQueueWrapper,
             applicationUserName: applicationUserName)
-        .whenComplete(
-            () => _sk1transactionObserver.cleanUpRestoredTransactions());
+        .whenComplete(() => _observer.cleanUpRestoredTransactions());
   }
 
   /// Query the product detail list.
