+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/types/app_store_purchase_details.dart
@@ -77,18 +77,3 @@ class AppStorePurchaseDetails extends PurchaseDetails {
   @override
   bool get pendingCompletePurchase => _pendingCompletePurchase;
 }
-
-/// The class represents the information of a purchase made with the Apple
-/// AppStore, when using Storekit2
-class SK2PurchaseDetails extends PurchaseDetails {
-  /// Creates new instance of [SK2PurchaseDetails]
-  SK2PurchaseDetails(
-      {required super.productID,
-      required super.purchaseID,
-      required super.verificationData,
-      required super.transactionDate,
-      required super.status});
-
-  @override
-  bool get pendingCompletePurchase => status == PurchaseStatus.purchased;
-}
