+++ packages/in_app_purchase/in_app_purchase_storekit/lib/store_kit_2_wrappers.dart
@@ -5,4 +5,3 @@
 export 'src/sk2_pigeon.g.dart';
 export 'src/store_kit_2_wrappers/sk2_appstore_wrapper.dart';
 export 'src/store_kit_2_wrappers/sk2_product_wrapper.dart';
-export 'src/store_kit_2_wrappers/sk2_transaction_wrapper.dart';
