+++ packages/in_app_purchase/in_app_purchase_storekit/test/fakes/fake_storekit_platform.dart
@@ -8,7 +8,6 @@ import 'package:in_app_purchase_storekit/in_app_purchase_storekit.dart';
 import 'package:in_app_purchase_storekit/src/messages.g.dart';
 import 'package:in_app_purchase_storekit/src/sk2_pigeon.g.dart';
 import 'package:in_app_purchase_storekit/src/store_kit_2_wrappers/sk2_product_wrapper.dart';
-import 'package:in_app_purchase_storekit/src/store_kit_2_wrappers/sk2_transaction_wrapper.dart';
 import 'package:in_app_purchase_storekit/store_kit_wrappers.dart';
 
 import '../sk2_test_api.g.dart';
@@ -285,13 +284,7 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
 class FakeStoreKit2Platform implements TestInAppPurchase2Api {
   late Set<String> validProductIDs;
   late Map<String, SK2Product> validProducts;
-  late List<SK2TransactionMessage> transactionList;
-  late bool testTransactionFail;
-  late int testTransactionCancel;
-  late List<SK2Transaction> finishedTransactions;
-
   PlatformException? queryProductException;
-  bool isListenerRegistered = false;
 
   void reset() {
     validProductIDs = <String>{'123', '456'};
@@ -334,50 +327,4 @@ class FakeStoreKit2Platform implements TestInAppPurchase2Api {
 
     return Future<List<SK2ProductMessage?>>.value(result);
   }
-
-  @override
-  Future<SK2ProductPurchaseResultMessage> purchase(String id,
-      {SK2ProductPurchaseOptionsMessage? options}) {
-    final SK2TransactionMessage transaction = createPendingTransaction(id);
-
-    InAppPurchaseStoreKitPlatform.sk2transactionObserver
-        .onTransactionsUpdated(transaction);
-    return Future<SK2ProductPurchaseResultMessage>.value(
-        SK2ProductPurchaseResultMessage.success);
-  }
-
-  @override
-  Future<void> finish(int id) {
-    return Future<void>.value();
-  }
-
-  @override
-  Future<List<SK2TransactionMessage?>> transactions() {
-    return Future<List<SK2TransactionMessage?>>.value(<SK2TransactionMessage?>[
-      SK2TransactionMessage(
-          id: 123,
-          originalId: 123,
-          productId: 'product_id',
-          purchaseDate: '12-12')
-    ]);
-  }
-
-  @override
-  void startListeningToTransactions() {
-    isListenerRegistered = true;
-  }
-
-  @override
-  void stopListeningToTransactions() {
-    isListenerRegistered = false;
-  }
-}
-
-SK2TransactionMessage createPendingTransaction(String id, {int quantity = 1}) {
-  return SK2TransactionMessage(
-      id: 1,
-      originalId: 2,
-      productId: id,
-      purchaseDate: 'purchaseDate',
-      appAccountToken: 'appAccountToken');
 }
