+++ packages/in_app_purchase/in_app_purchase_storekit/test/in_app_purchase_storekit_2_platform_test.dart
@@ -2,27 +2,15 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:async';
-
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:in_app_purchase_platform_interface/in_app_purchase_platform_interface.dart';
 import 'package:in_app_purchase_storekit/in_app_purchase_storekit.dart';
-import 'package:in_app_purchase_storekit/store_kit_2_wrappers.dart';
 
 import 'fakes/fake_storekit_platform.dart';
 import 'sk2_test_api.g.dart';
 
 void main() {
-  final SK2Product dummyProductWrapper = SK2Product(
-      id: '2',
-      displayName: 'name',
-      displayPrice: '0.99',
-      description: 'desc',
-      price: 0.99,
-      type: SK2ProductType.consumable,
-      priceLocale: SK2PriceLocale(currencyCode: 'USD', currencySymbol: r'$'));
-
   TestWidgetsFlutterBinding.ensureInitialized();
 
   final FakeStoreKit2Platform fakeStoreKit2Platform = FakeStoreKit2Platform();
@@ -80,75 +68,6 @@ void main() {
       expect(response.error!.code, 'error_code');
       expect(response.error!.message, 'error_message');
       expect(response.error!.details, <Object, Object>{'info': 'error_info'});
-    });
-  });
-
-  group('make payment', () {
-    test(
-        'buying non consumable, should get purchase objects in the purchase update callback',
-        () async {
-      final List<PurchaseDetails> details = <PurchaseDetails>[];
-      final Completer<List<PurchaseDetails>> completer =
-          Completer<List<PurchaseDetails>>();
-      final Stream<List<PurchaseDetails>> stream =
-          iapStoreKitPlatform.purchaseStream;
-
-      late StreamSubscription<List<PurchaseDetails>> subscription;
-      subscription = stream.listen((List<PurchaseDetails> purchaseDetailsList) {
-        details.addAll(purchaseDetailsList);
-        if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
-          completer.complete(details);
-          subscription.cancel();
-        }
-      });
-      final AppStorePurchaseParam purchaseParam = AppStorePurchaseParam(
-          productDetails:
-              AppStoreProduct2Details.fromSK2Product(dummyProductWrapper),
-          applicationUserName: 'appName');
-      await iapStoreKitPlatform.buyNonConsumable(purchaseParam: purchaseParam);
-
-      final List<PurchaseDetails> result = await completer.future;
-      expect(result.length, 1);
-      expect(result.first.productID, dummyProductWrapper.id);
-    });
-
-    test(
-        'buying consumable, should get purchase objects in the purchase update callback',
-        () async {
-      final List<PurchaseDetails> details = <PurchaseDetails>[];
-      final Completer<List<PurchaseDetails>> completer =
-          Completer<List<PurchaseDetails>>();
-      final Stream<List<PurchaseDetails>> stream =
-          iapStoreKitPlatform.purchaseStream;
-
-      late StreamSubscription<List<PurchaseDetails>> subscription;
-      subscription = stream.listen((List<PurchaseDetails> purchaseDetailsList) {
-        details.addAll(purchaseDetailsList);
-        if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
-          completer.complete(details);
-          subscription.cancel();
-        }
-      });
-      final AppStorePurchaseParam purchaseParam = AppStorePurchaseParam(
-          productDetails:
-              AppStoreProduct2Details.fromSK2Product(dummyProductWrapper),
-          applicationUserName: 'appName');
-      await iapStoreKitPlatform.buyConsumable(purchaseParam: purchaseParam);
-
-      final List<PurchaseDetails> result = await completer.future;
-      expect(result.length, 1);
-      expect(result.first.productID, dummyProductWrapper.id);
-    });
-
-    test('buying consumable, should throw when autoConsume is false', () async {
-      final AppStorePurchaseParam purchaseParam = AppStorePurchaseParam(
-          productDetails:
-              AppStoreProduct2Details.fromSK2Product(dummyProductWrapper),
-          applicationUserName: 'appName');
-      expect(
-          () => iapStoreKitPlatform.buyConsumable(
-              purchaseParam: purchaseParam, autoConsume: false),
-          throwsA(isInstanceOf<AssertionError>()));
     });
   });
 }
